pyproject.toml
[project]
name = "nanovlm"
package-mode = false
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "==3.11.12"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "huggingface-hub (>=0.32.0,<0.33.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "gcsfs (==2025.3.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "transformers (>=4.52.3,<5.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
